requirements.txt
langchain
langchain-google-genai
pypdf2
python-dotenv
google-generativeai
langchain-community
openai